Commit2

+++ app/Http/Controllers/UsersController.php
@@ -4,6 +4,8 @@ namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
 use App\Models\User;
+use Illuminate\Support\Facades\Hash;
+
 
 class UsersController extends Controller
 {
@@ -19,20 +21,36 @@ class UsersController extends Controller
 
     //crear un usuario
     public function store(Request $request){
-        return User::create($request->all());
+        $this->validate($request, [
+            'username' => 'required|unique:users',
+            'password' => 'required',
+            'rol' => 'required',
+
+        ]);
+        $user = new User;
+        $user->fill($request->all());
+        $user->password= Hash::make($request->password);
+        $user->save();
+        return $user;
 
 }
 
     //actualizar
     public function update(Request $request, $id){
+        $this->validate($request, [
+            'username' => 'filled|unique:users',
+        ]);
         $user = User::find($id);
+        if(!$user) return response('', 404);
         $user->update($request->all());
+        if($request->password) $user->password= Hash::make($request->password);
         $user->save();
         return $user;
     }
 
     public function destroy($id){
         $user = User::find($id);
+        if(!$user) return response('', 404);
         $user->delete();
         return $user;
     }
